requirements.txt
# Core ML and NLP libraries
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.3.0

# Vector databases
faiss-cpu>=1.7.0
pinecone>=5.0.0

# Text processing and search
nltk>=3.7
spacy>=3.4.0
rank-bm25>=0.2.2

# Data handling and utilities
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
jupyter>=1.0.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=0.19.0
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Optional: For advanced legal text processing
# legal-bert (if available via pip)
# spacy-transformers>=1.1.0
